Added checks if files are existing before deleting

+++ devops.sh
@@ -3,8 +3,14 @@
 bin_dir="$(linked_file=$(readlink ${0}) && cd $(cd $(dirname ${0}) && echo $(pwd)) && cd $(dirname ${linked_file}) && echo $(pwd))"
 includes_dir="${bin_dir}/includes"
 
-rm "$(pwd)/server.sh"
+if [ -f "$(pwd)/server.sh" ]; then
+    rm "$(pwd)/server.sh"
+fi
+
 ln -s "${bin_dir}/server.sh" "$(pwd)/server.sh"
 
-rm "$(pwd)/wp.sh"
+if [ -f "$(pwd)/wp.sh" ]; then
+    rm "$(pwd)/wp.sh"
+fi
+
 ln -s "${bin_dir}/wp.sh" "$(pwd)/wp.sh"
